renamed package command to formdata
