+++ worblehat-web/src/main/java/de/codecentric/psd/worblehat/web/controller/BorrowBookController.java
@@ -3,7 +3,7 @@ package de.codecentric.psd.worblehat.web.controller;
 import de.codecentric.psd.worblehat.domain.Book;
 import de.codecentric.psd.worblehat.domain.BookAlreadyBorrowedException;
 import de.codecentric.psd.worblehat.domain.BookService;
-import de.codecentric.psd.worblehat.web.command.BookBorrowFormData;
+import de.codecentric.psd.worblehat.web.formdata.BookBorrowFormData;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
