+++ worblehat-web/src/main/java/de/codecentric/psd/worblehat/web/formdata/ReturnAllBooksFormData.java
@@ -1,0 +1,23 @@
+package de.codecentric.psd.worblehat.web.formdata;
+
+import org.hibernate.validator.constraints.Email;
+import org.hibernate.validator.constraints.NotEmpty;
+
+/**
+ * This class represent the form data of the return book form.
+ */
+public class ReturnAllBooksFormData {
+
+	@NotEmpty(message = "{empty.returnAllBookFormData.emailAddress}") 
+	@Email(message = "{notvalid.returnAllBookFormData.emailAddress}")
+	private String emailAddress;
+
+	public String getEmailAddress() {
+		return emailAddress;
+	}
+
+	public void setEmailAddress(String emailAddress) {
+		this.emailAddress = emailAddress;
+	}
+
+}
